ENG-2094 Hide redis table and keyspace in YQL.
Summary:
redis_keyspace and the .redis table show up in cqlsh and querying them lead to tserver
crashes. We shouldn't really expose any redis specific stuff to the YQL layer. This diff ensures
that SELECT, INSERT, UPDATE, CREATE and DROP return appropriate errors for redis_keyspace and
.redis.

It also ensures that redis_keyspace and .redis table are not displayed in describe keyspaces or describe tables. We also renamed 'redis_keyspace' to 'system_redis'.

Test Plan:
1) jenkins
2) Unit test added.

Reviewers: robert, neil, bogdan, oleg

Reviewed By: oleg

Subscribers: karthik, ybase, kannan, sid

Differential Revision: https://phabricator.dev.yugabyte.com/D2982

+++ java/yb-cql/src/test/java/org/yb/cql/BaseCQLTest.java
@@ -33,6 +33,7 @@ import com.datastax.driver.core.exceptions.QueryValidationException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import org.yb.client.YBClient;
 import org.yb.master.Master;
 import org.yb.minicluster.BaseMiniClusterTest;
 import org.yb.minicluster.IOMetrics;
@@ -299,6 +300,10 @@ public class BaseCQLTest extends BaseMiniClusterTest {
            keyspaceName.equals("system_schema");
   }
 
+  private boolean IsRedisKeyspace(String keyspaceName) {
+    return keyspaceName.equals(YBClient.REDIS_KEYSPACE_NAME);
+  }
+
   protected void dropTables() throws Exception {
     if (miniCluster == null) {
       return;
@@ -307,7 +312,7 @@ public class BaseCQLTest extends BaseMiniClusterTest {
         miniCluster.getClient().getTablesList().getTableInfoList()) {
       // Drop all non-system tables.
       String namespaceName = tableInfo.getNamespace().getName();
-      if (!IsSystemKeyspace(namespaceName)) {
+      if (!IsSystemKeyspace(namespaceName) && !IsRedisKeyspace(namespaceName)) {
         dropTable(namespaceName + "." + tableInfo.getName());
       }
     }
